Adds the scaffolding for http codes
Adds the beginning scaffolding for https codes

These will have a code, name, description, category. Code refers to the
three digit number that is tradionally thought of. Name is the short
named version of the https code. Description will have a brief
description of how the code is used. Category will refer to one of the
five categories of https codes

This commit includes most importantly the migration described above, but
also the other basics that are generated.